Make crossover and mutation dependent on variation score.
Doing this makes the algorithm hit its peak (of 0.75) more often. Rigging
mutation to mutate at most n times (where n is the number of pegs not judged to
be in the right place, misplaced and completely wrong alike) makes the score
less likely to drop back to 0, for some reason.